refactor(browser): move tag editing to the ViewModel
